Migrate linked figures
