+++ config/ignores/index.js
@@ -1,0 +1,5 @@
+const ignoresPlugin = (eleventyConfig) => {
+  eleventyConfig.watchIgnores.add('.cache/related-posts-*.md')
+}
+
+module.exports = ignoresPlugin;
