+++ config/index.js
@@ -5,6 +5,7 @@ const data = require('./data')
 const dataExtensions = require('./data_extensions')
 const events = require('./events')
 const filters = require('./filters')
+const ignores = require('./ignores')
 const markdown = require('./markdown')
 const passthroughs = require('./passthroughs')
 const plugins = require('./plugins')
@@ -20,6 +21,7 @@ module.exports = function EighteenF(eleventyConfig) {
   eleventyConfig.addPlugin(dataExtensions)
   eleventyConfig.addPlugin(events)
   eleventyConfig.addPlugin(filters)
+  eleventyConfig.addPlugin(ignores)
   eleventyConfig.addPlugin(markdown)
   eleventyConfig.addPlugin(passthroughs)
   eleventyConfig.addPlugin(plugins)
