+++ content/posts/2020-06-30-transforming-how-dojs-civil-rights-division-engages-with-the-public.md
@@ -26,12 +26,10 @@ image: /assets/blog/blog-post_header-2.png
 ---
 On Feb. 7, 1961, James Meredith submitted a [letter](https://www.jfklibrary.org/asset-viewer/archives/BMPP/020/BMPP-020-007) to the Department of Justice (DOJ) asking for the Department’s support in his admission to the University of Mississippi by enforcing federal desegregation laws not being respected by the state. Through the combined efforts of the local NAACP, the Department of Justice, and the US Supreme Court, Meredith eventually became the first African-American student admitted to the university.
 
-<figure>
-  <a href="{{site.baseurl}}/assets/blog/james-meredith-letter-sm.png">
-    <img src="{{site.baseurl}}/assets/blog/james-meredith-letter-sm.png" alt="Image of James Meredith letter submitted to the Department of Justice (DOJ) asking for the Department’s support in his admission to the University of Mississippi by enforcing federal desegregation laws not being respected by the state."/>
-  </a>
-  <figcaption>James Meredith letter submitted to the Department of Justice (DOJ)</figcaption>
-</figure>
+{% include "linked-figure.html",
+     image: "/assets/blog/james-meredith-letter-sm.png",
+     alt: "Image of James Meredith letter submitted to the Department of Justice (DOJ) asking for the Department’s support in his admission to the University of Mississippi by enforcing federal desegregation laws not being respected by the state.",
+     caption: "James Meredith letter submitted to the Department of Justice (DOJ)" %}
 
 Since 1961, the Department of Justice has come a long way in making it easier for people to report discrimination or a violation of rights. Now, with the help of 18F, they have transformed how the Civil Rights Division receives and processes civil rights reports using modern technology and user experience design practices.
 
@@ -44,23 +42,19 @@ Today, the Civil Rights Division estimates that they receive over 100,000 report
 
 The Division also receives reports of potential voting rights violations, particularly during election season.
 
-<figure>
-  <a href="{{site.baseurl}}/assets/blog/how-to-file_old.gif">
-    <img src="{{site.baseurl}}/assets/blog/how-to-file_old.gif" alt="Previous “how to file a complaint site."/>
-  </a>
-  <figcaption>Previous “how to file a complaint” experience</figcaption>
-</figure>
+{% include "linked-figure.html",
+     image: "/assets/blog/how-to-file_old.gif",
+     alt: "Previous “how to file a complaint” site.",
+     caption: "Previous “how to file a complaint” experience" %}
 
 Up until now, the experience of filing a report was confusing and inconsistent for the public. The Civil Rights Division is a complex organization with several specialized units (or “sections”). This complexity was confusing for some victims of civil rights violations. A victim reporting a violation had to pick among dozens of possible reporting pathways. Without clear guidance and expectation setting, many people submitted reports that required extensive follow-up to collect missing information, or submitted reports that fell outside the purview of the Division altogether.
 
 In order to be more responsive to the public’s changing communication needs and the increased reporting volume,the Civil Rights Division, in close collaboration with 18F, [has launched](https://www.justice.gov/opa/pr/department-justice-announces-launch-civil-rights-reporting-portal) a user-friendly online submission experience at [civilrights.justice.gov](https://civilrights.justice.gov/) that transforms the way the Division collects, sorts, and responds to civil rights reports.
 
-<figure>
-  <a href="{{site.baseurl}}/assets/blog/doj-crt-landing-page.gif">
-    <img src="{{site.baseurl}}/assets/blog/doj-crt-landing-page.gif" alt="The new Civil Rights Division landing page of civilright.justice.gov."/>
-  </a>
-  <figcaption>The new experience for reporting a civil rights violation</figcaption>
-</figure>
+{% include "linked-figure.html",
+     image: "/assets/blog/doj-crt-landing-page.gif",
+     alt: "The new Civil Rights Division landing page of civilright.justice.gov.",
+     caption: "The new experience for reporting a civil rights violation" %}
 
 We worked with stakeholders from the Division to craft these design principles for the new experience, which subsequent designs, prototypes, and builds were tested against:
 
@@ -73,12 +67,10 @@ Backed with in-depth usability research, this new user-centered online reporting
 
 For Division staff, this new experience creates a more consistent, collaborative, and transparent process. It allows staff to easily review, sort, and redirect reports to the appropriate team for analysis. This saves time and energy, freeing up more capacity for the Division to focus on advancing civil rights.
 
-<figure>
-  <a href="{{site.baseurl}}/assets/blog/doj-crt-complaint-form.png">
-    <img src="{{site.baseurl}}/assets/blog/doj-crt-complaint-form.png" alt="Pages of the new online civil rights complaint form."/>
-  </a>
-  <figcaption>Pages of the new online civil rights complaint form</figcaption>
-</figure>
+{% include "linked-figure.html",
+     image: "/assets/blog/doj-crt-complaint-form.png",
+     alt: "Pages of the new online civil rights complaint form.",
+     caption: "Pages of the new online civil rights complaint form" %}
 
 The core development team across the Civil Rights Division and 18F is still iterating on the experience. After launching for internal use and as a ‘beta’ experience, the team is announcing this website more widely so we can continue to improve and adapt with more real reports and feedback from the general public.
